Refactor jndi services to reduce package cycles.
