+++ project-set/core/core-lib/src/main/java/com/rackspace/papi/service/context/FilterChainGCServiceContext.java
@@ -1,4 +1,4 @@
-package com.rackspace.papi.service.filterchain;
+package com.rackspace.papi.service.context;
 
 import com.rackspace.papi.commons.util.thread.DestroyableThreadWrapper;
 import com.rackspace.papi.commons.util.thread.Poller;
@@ -6,6 +6,7 @@ import com.rackspace.papi.commons.util.thread.RecurringTask;
 import com.rackspace.papi.filter.resource.SweepingGarbageCollector;
 import com.rackspace.papi.service.ServiceContext;
 import com.rackspace.papi.service.context.jndi.ServletContextHelper;
+import com.rackspace.papi.service.filterchain.GarbageCollectionService;
 import com.rackspace.papi.service.threading.ThreadingService;
 import javax.servlet.ServletContextEvent;
