+++ project-set/core/core-lib/src/main/java/com/rackspace/papi/service/context/LoggingServiceContext.java
@@ -1,13 +1,17 @@
-package com.rackspace.papi.service.logging;
+package com.rackspace.papi.service.context;
 
 import com.rackspace.papi.commons.config.manager.UpdateListener;
 import com.rackspace.papi.commons.util.StringUtilities;
-import com.rackspace.papi.container.config.ContainerConfiguration;
 
+import com.rackspace.papi.container.config.ContainerConfiguration;
 import com.rackspace.papi.container.config.LoggingConfiguration;
+
 import com.rackspace.papi.service.ServiceContext;
 import com.rackspace.papi.service.config.ConfigurationService;
 import com.rackspace.papi.service.context.jndi.ServletContextHelper;
+import com.rackspace.papi.service.logging.LoggingService;
+import com.rackspace.papi.service.logging.LoggingServiceImpl;
+import com.rackspace.papi.service.logging.common.LogFrameworks;
 
 import javax.servlet.ServletContext;
 import javax.servlet.ServletContextEvent;
