+++ project-set/core/core-lib/src/main/java/com/rackspace/papi/service/context/jndi/JndiContextAdapter.java
@@ -2,19 +2,19 @@ package com.rackspace.papi.service.context.jndi;
 
 import com.rackspace.papi.service.ServiceContext;
 import com.rackspace.papi.service.config.ConfigurationService;
+import com.rackspace.papi.service.context.ClassLoaderServiceContext;
+import com.rackspace.papi.service.context.LoggingServiceContext;
 import com.rackspace.papi.service.datastore.DatastoreService;
 import com.rackspace.papi.service.logging.LoggingService;
-import com.rackspace.papi.service.logging.LoggingServiceContext;
 import com.rackspace.papi.service.rms.ResponseMessageService;
 import com.rackspace.papi.service.ServiceUnavailableException;
 import com.rackspace.papi.service.event.EventService;
 import com.rackspace.papi.service.event.EventManagerServiceContext;
 import com.rackspace.papi.service.classloader.ClassLoaderManagerService;
-import com.rackspace.papi.service.classloader.ClassLoaderServiceContext;
-import com.rackspace.papi.service.config.ConfigurationServiceContext;
+import com.rackspace.papi.service.context.ConfigurationServiceContext;
 import com.rackspace.papi.service.context.ResponseMessageServiceContext;
 import com.rackspace.papi.service.datastore.DatastoreServiceContext;
-import com.rackspace.papi.service.filterchain.FilterChainGCServiceContext;
+import com.rackspace.papi.service.context.FilterChainGCServiceContext;
 import com.rackspace.papi.service.filterchain.GarbageCollectionService;
 import com.rackspace.papi.service.threading.ThreadingServiceContext;
 import com.rackspace.papi.service.threading.ThreadingService;
